Moved `@Rollback` to its own module #24500
